requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-admin-interface==0.28.1
django-colorfield==0.11.0
django-debug-toolbar==4.2.0
django-email-users==1.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
drf-writable-nested==0.1.0
flake8==7.0.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
routers==0.10.1
rpds-py==0.16.2
slugify==0.0.1
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.1.0
